# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Investigation & Policing Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suspects:
    get:
      operationId: listSuspects
      tags: [suspects]
      summary: List suspects
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: riskLevel
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of suspects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Suspect"
    post:
      operationId: createSuspect
      tags: [suspects]
      summary: Add a suspect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspectInput"
      responses:
        "201":
          description: Suspect created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suspect"
  /suspects/{id}:
    get:
      operationId: getSuspect
      tags: [suspects]
      summary: Get suspect details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspect details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suspect"
        "404":
          description: Not found
    patch:
      operationId: updateSuspect
      tags: [suspects]
      summary: Update suspect record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspectUpdate"
      responses:
        "200":
          description: Updated suspect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suspect"
...
components:
  schemas:
...
    Suspect:
      type: object
      required: [id, name, alias, riskLevel, status, knownAssociates, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        alias: { type: ["string", "null"] }
        riskLevel: { type: string }
        status: { type: string }
        nationalId: { type: ["string", "null"] }
        knownAssociates: { type: ["string", "null"] }
        socialMediaProfiles: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        caseIds: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    SuspectInput:
      type: object
      required: [name, riskLevel]
      properties:
        name: { type: string }
        alias: { type: string }
        riskLevel: { type: string }
        nationalId: { type: string }
        socialMediaProfiles: { type: string }
        notes: { type: string }
    SuspectUpdate:
      type: object
      properties:
        name: { type: string }
        alias: { type: string }
        riskLevel: { type: string }
        status: { type: string }
        knownAssociates: { type: string }
        socialMediaProfiles: { type: string }
        notes: { type: string }
...