# lib/api-zod/src/generated/types/suspectUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Investigation & Policing Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface SuspectUpdate {
  name?: string;
  alias?: string;
  riskLevel?: string;
  status?: string;
  knownAssociates?: string;
  socialMediaProfiles?: string;
  notes?: string;
}